requirements.txt
pyautogui
pyperclip
keyboard
PyYAML
python-dotenv
google-generativeai
Pillow